Initial setup with video components
